+++ src/c++/lib/format/VcfWriterSV.cpp
@@ -417,10 +417,14 @@ writeTransloc(
     {
         if (bpArange.size() > 1)
         {
+            // get breakend homology sequence
             infotags.push_back( str( boost::format("HOMLEN=%i") % (bpArange.size()-1) ));
-            std::string homref;
-            get_standardized_region_seq(_referenceFilename,chrom,bpArange.begin_pos()+1,bpArange.end_pos()-1,homref);
-            infotags.push_back( str( boost::format("HOMSEQ=%s") % (homref) ));
+            std::string homSeq;
+            // the get region seq function below takes closed-closed, 0-based endpoints
+            const int homBegin(bpArange.begin_pos());
+            const int homEnd(bpArange.end_pos()-2);
+            get_standardized_region_seq(_referenceFilename,chrom,homBegin,homEnd,homSeq);
+            infotags.push_back( str( boost::format("HOMSEQ=%s") % (homSeq) ));
         }
     }
 
@@ -654,10 +658,14 @@ writeInvdel(
     {
         if (bpArange.size() > 1)
         {
+            // get breakend homology sequence
             infoTags.push_back( str( boost::format("HOMLEN=%i") % (bpArange.size()-1) ));
-            std::string homref;
-            get_standardized_region_seq(_referenceFilename,chrom,bpArange.begin_pos()+1,bpArange.end_pos()-1,homref);
-            infoTags.push_back( str( boost::format("HOMSEQ=%s") % (homref) ));
+            std::string homSeq;
+            // the get region seq function below takes closed-closed, 0-based endpoints
+            const int homBegin(bpArange.begin_pos());
+            const int homEnd(bpArange.end_pos()-2);
+            get_standardized_region_seq(_referenceFilename,chrom,homBegin,homEnd,homSeq);
+            infoTags.push_back( str( boost::format("HOMSEQ=%s") % (homSeq) ));
         }
     }